Improve login and register UI
Update register screen's design to match the general design of the application, and enable the progress bar.

Add a proper handling of the registration results.

Improve the logo placement on the login screen.

Improve management of the UI with the definition of styles which :
	- Correct the colors of the login screen
	- Set a default style for buttons and text fields

Switch type of fields from EditText to more correct TextInputEditText.